+++ src/components/mbti_meeting/MbtiMeeting.jsx
@@ -61,7 +61,7 @@ const MbtiMeeting = () => {
         });
         setMeet(initialMeet);
     };
-    console.log('진재석');
+
     // 위로 올라가기 버튼
     useEffect(() => {
         const scroll = () => {
