+++ src/recoil/recoilAtoms.js
@@ -8,57 +8,13 @@ export const tagCategoryState = atom({
     default: ''
 });
 
+// MBTI Meeting 메인 페이지 검색 정보 저장
 export const selectedTagsState = atom({
     key: 'selectedTagsState',
     default: []
 });
 
-// 모임 정보
-export const meetingRepreImgState = atom({
-    key: 'meetingRepreImgState',
-    default: ''
-});
-
-export const meetingNameState = atom({
-    key: 'meetingNameState',
-    default: ''
-});
-
-export const meetingManagerNameState = atom({
-    key: 'meetingManagerNameState',
-    default: ''
-});
-
-export const meetingLimitPeopleState = atom({
-    key: 'meetingLimitPeopleState',
-    default: ''
-});
-
-export const meetingScheduleState = atom({
-    key: 'meetingScheduleState',
-    default: ''
-});
-
-export const meetingKakaoUrlState = atom({
-    key: 'meetingKakaoUrlState',
-    default: ''
-});
-
-export const meetingOneLineIntroState = atom({
-    key: 'meetingOneLineIntroState',
-    default: ''
-});
-
-export const meetingIntroTitleState = atom({
-    key: 'meetingIntroTitleState',
-    default: ''
-});
-
-export const meetingIntroContentState = atom({
-    key: 'meetingIntroContentState',
-    default: ''
-});
-
+// MBTI Meeting 생성 페이지 입력 정보 저장
 export const createMeetingState = atom({
     key: 'newMeeting',
     default: {}
